generate metadata in the hook
